+++ packages/regal/lib/src/theme/theme_data.dart
@@ -182,7 +182,7 @@ final kThemeData = ThemeData.light().copyWith(
       ),
       labelSmall: kFigTreeTextTheme.labelSmall!.copyWith(
         color: RegalColors.licoriceBlack,
-        fontSize: 16.sp,
+        fontSize: 13.sp,
         fontWeight: FontWeight.w400,
       ),
       bodyLarge: kFigTreeTextTheme.bodyLarge!.copyWith(
@@ -577,7 +577,6 @@ final kDarkThemeData = kThemeData.copyWith(
       titleMedium: kFigTreeTextTheme.titleMedium!.copyWith(
         fontSize: 20.sp,
         color: RegalColors.snowWhite,
-        fontFamily: 'Montserrat',
       ),
       titleSmall: kFigTreeTextTheme.titleSmall!.copyWith(
         color: RegalColors.snowWhite,
@@ -596,7 +595,7 @@ final kDarkThemeData = kThemeData.copyWith(
       ),
       labelSmall: kFigTreeTextTheme.labelSmall!.copyWith(
         color: RegalColors.snowWhite,
-        fontSize: 16.sp,
+        fontSize: 13.sp,
         fontWeight: FontWeight.w400,
       ),
       bodyLarge: kFigTreeTextTheme.bodyLarge!.copyWith(